# openapi.json
{"openapi":"3.0.0","paths":{"/claim-tokens":{"post":{"operationId":"ClaimTokensController_generateClaimToken","parameters":[],"responses":{"201":{"description":""}},"tags":["ClaimTokens"]}},"/claim-tokens/{token}":{"get":{"operationId":"ClaimTokensController_verifyClaimToken","parameters":[{"name":"token","required":true,"in":"path","description":"Claim token string","schema":{"type":"string"}}],"responses":{"200":{"description":"Claim token verified."}},"summary":"Verify a claim token","tags":["ClaimTokens"]}},"/claim-tokens/claim":{"post":{"operationId":"ClaimTokensController_claimDevice","parameters":[],"responses":{"201":{"description":""}},"tags":["ClaimTokens"]}},"/apikeys":{"post":{"operationId":"ApiKeysController_createApiKey","parameters":[],"requestBody":{"required":true,"description":"Data for creating the API key","content":{"application/json":{"schema":{"type":"object","properties":{"description":{"type":"string","example":"Mobile Application"},"role":{"type":"string","example":"admin"}},"required":["description","role"]}}}},"responses":{"201":{"description":"API key successfully created."}},"security":[{"x-api-key":[]}],"summary":"Create a new API key","tags":["API Keys"]},"get":{"operationId":"ApiKeysController_getAllApiKeys","parameters":[],"responses":{"200":{"description":"List of API keys."}},"security":[{"x-api-key":[]}],"summary":"List all registered API keys","tags":["API Keys"]}},"/apikeys/{id}":{"get":{"operationId":"ApiKeysController_getApiKeyById","parameters":[{"name":"id","required":true,"in":"path","description":"API key ID","schema":{"type":"string"}}],"responses":{"200":{"description":"API key details."}},"security":[{"x-api-key":[]}],"summary":"Get details of a specific API key","tags":["API Keys"]},"put":{"operationId":"ApiKeysController_updateApiKey","parameters":[{"name":"id","required":true,"in":"path","description":"API key ID","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Data for updating the API key","content":{"application/json":{"schema":{"type":"object","properties":{"description":{"type":"string"},"active":{"type":"boolean"}}}}}},"responses":{"200":{"description":"API key updated."}},"security":[{"x-api-key":[]}],"summary":"Update an API key","tags":["API Keys"]}},"/mqtt/{deviceSerial}/control":{"post":{"operationId":"MqttController_sendControlCommand","parameters":[{"name":"deviceSerial","required":true,"in":"path","description":"Device serial","schema":{"example":"ABC123","type":"string"}}],"requestBody":{"required":true,"description":"Control command payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Object"}}}},"responses":{"201":{"description":"Control command sent successfully."}},"security":[{"x-api-key":[]}],"summary":"Send control command to the device","tags":["MQTT"]}},"/mqtt/{deviceSerial}/management":{"post":{"operationId":"MqttController_sendManagementCommand","parameters":[{"name":"deviceSerial","required":true,"in":"path","description":"Device serial","schema":{"example":"ABC123","type":"string"}}],"requestBody":{"required":true,"description":"Management command payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Object"}}}},"responses":{"201":{"description":"Management command sent successfully."}},"security":[{"x-api-key":[]}],"summary":"Send management command to the device","tags":["MQTT"]}},"/events":{"get":{"operationId":"EventsController_getFilteredEvents","parameters":[{"name":"deviceSerial","required":true,"in":"query","schema":{"type":"string"}},{"name":"epcPrefix","required":true,"in":"query","schema":{"type":"string"}},{"name":"antenna","required":true,"in":"query","schema":{"type":"string"}},{"name":"rssiMin","required":true,"in":"query","schema":{"type":"string"}},{"name":"rssiMax","required":true,"in":"query","schema":{"type":"string"}},{"name":"from","required":true,"in":"query","schema":{"type":"string"}},{"name":"to","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"x-api-key":[]}],"tags":["Events"]}},"/commands":{"get":{"operationId":"CommandsController_getCommands","parameters":[{"name":"to","required":false,"in":"query","description":"End date for filtering (ISO string)","schema":{}},{"name":"from","required":false,"in":"query","description":"Start date for filtering (ISO string)","schema":{}},{"name":"deviceSerial","required":false,"in":"query","description":"Filter by device serial","schema":{}},{"name":"status","required":false,"in":"query","description":"Filter by command status (pending, success, error)","schema":{}},{"name":"type","required":false,"in":"query","description":"Filter by command type (control/management)","schema":{}}],"responses":{"200":{"description":"List of commands."}},"security":[{"x-api-key":[]}],"summary":"Get command history","tags":["Commands"]},"post":{"operationId":"CommandsController_createCommand","parameters":[],"responses":{"201":{"description":""}},"security":[{"x-api-key":[]}],"tags":["Commands"]}},"/commands/{commandId}":{"get":{"operationId":"CommandsController_getCommandById","parameters":[{"name":"commandId","required":true,"in":"path","description":"Command ID","schema":{"example":"123e4567-e89b-12d3-a456-426614174000","type":"string"}}],"responses":{"200":{"description":"Command details."}},"security":[{"x-api-key":[]}],"summary":"Get details of a specific command","tags":["Commands"]}},"/commands/{commandId}/status":{"get":{"operationId":"CommandsController_getCommandStatus","parameters":[{"name":"commandId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"x-api-key":[]}],"tags":["Commands"]}},"/metrics/reader":{"get":{"operationId":"MetricsController_getReaderMetrics","parameters":[{"name":"to","required":false,"in":"query","description":"End date (ISO)","schema":{}},{"name":"from","required":false,"in":"query","description":"Start date (ISO)","schema":{}},{"name":"deviceSerial","required":true,"in":"query","description":"Device serial","schema":{}}],"responses":{"200":{"description":"RFID reader metrics"}},"security":[{"x-api-key":[]}],"summary":"Get RFID reader metrics","tags":["Metrics"]}},"/metrics/antennas":{"get":{"operationId":"MetricsController_getAntennaMetrics","parameters":[{"name":"to","required":false,"in":"query","description":"End date (ISO)","schema":{}},{"name":"from","required":false,"in":"query","description":"Start date (ISO)","schema":{}},{"name":"deviceSerial","required":true,"in":"query","description":"Device serial","schema":{}}],"responses":{"200":{"description":"Antenna metrics"}},"security":[{"x-api-key":[]}],"summary":"Get RFID reader antenna metrics","tags":["Metrics"]}},"/metrics/system":{"get":{"operationId":"MetricsController_getSystemMetrics","parameters":[{"name":"to","required":false,"in":"query","description":"End date (ISO)","schema":{}},{"name":"from","required":false,"in":"query","description":"Start date (ISO)","schema":{}},{"name":"deviceSerial","required":true,"in":"query","description":"Device serial","schema":{}}],"responses":{"200":{"description":"System metrics"}},"security":[{"x-api-key":[]}],"summary":"Get system metrics (CPU, memory)","tags":["Metrics"]}},"/metrics/offline":{"get":{"operationId":"MetricsController_getOfflineDevices","parameters":[{"name":"minutes","required":true,"in":"query","description":"Time interval in minutes to consider the device offline","schema":{"type":"string"}}],"responses":{"200":{"description":"List of offline devices with the last communication date"}},"security":[{"x-api-key":[]}],"summary":"Return devices without communication for more than X minutes","tags":["Metrics"]}},"/metrics/history":{"get":{"operationId":"MetricsController_getMetricsHistory","parameters":[{"name":"from","required":false,"in":"query","description":"Start date for filtering metrics (ISO 8601 format)","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","description":"End date for filtering metrics (ISO 8601 format)","schema":{"type":"string"}}],"responses":{"200":{"description":"List of system metrics within the specified time range"}},"security":[{"x-api-key":[]}],"summary":"Retrieve system metrics history within a given time range","tags":["Metrics"]}},"/metrics":{"get":{"operationId":"MetricsController_getMetrics","parameters":[],"responses":{"200":{"description":"Prometheus-formatted metrics"}},"security":[{"x-api-key":[]}],"summary":"Retrieve Prometheus system metrics","tags":["Metrics"]}},"/monitoring/dashboard":{"get":{"operationId":"MonitoringController_getDashboardData","parameters":[{"name":"deviceSerial","required":false,"in":"query","description":"Optional filter by deviceSerial","schema":{}},{"name":"to","required":true,"in":"query","description":"End date (ISO)","schema":{}},{"name":"from","required":true,"in":"query","description":"Start date (ISO)","schema":{}}],"responses":{"200":{"description":"Availability data for the dashboard"}},"security":[{"x-api-key":[]}],"summary":"Return availability data for the dashboard","tags":["Monitoring Dashboard"]}},"/api/devices/{deviceSerial}/control":{"post":{"operationId":"DevicesController_sendControlCommand","parameters":[{"name":"deviceSerial","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"x-api-key":[]}],"summary":"Send control command to a device","tags":["Devices"]}},"/api/devices":{"post":{"operationId":"DevicesController_createDevice","parameters":[],"responses":{"201":{"description":""}},"security":[{"x-api-key":[]}],"summary":"Register a new device","tags":["Devices"]},"get":{"operationId":"DevicesController_getAllDevices","parameters":[],"responses":{"200":{"description":""}},"security":[{"x-api-key":[]}],"summary":"List all devices","tags":["Devices"]}},"/api/devices/{deviceId}":{"put":{"operationId":"DevicesController_updateDevice","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"x-api-key":[]}],"summary":"Update device information","tags":["Devices"]},"delete":{"operationId":"DevicesController_deleteDevice","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"x-api-key":[]}],"summary":"Delete a device","tags":["Devices"]}},"/api/devices/{deviceSerial}/firmware":{"patch":{"operationId":"DevicesController_updateFirmwareVersion","parameters":[{"name":"deviceSerial","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"x-api-key":[]}],"tags":["Devices"]}},"/api/devices/{deviceSerial}/configuration":{"patch":{"operationId":"DevicesController_updateDeviceConfiguration","parameters":[{"name":"deviceSerial","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"x-api-key":[]}],"tags":["Devices"]}},"/webhooks":{"post":{"operationId":"WebhooksController_subscribe","parameters":[],"responses":{"201":{"description":""}},"tags":["Webhooks"]},"get":{"operationId":"WebhooksController_getSubscriptions","parameters":[],"responses":{"200":{"description":""}},"tags":["Webhooks"]},"delete":{"operationId":"WebhooksController_deleteSubscription","parameters":[{"name":"url","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Webhooks"]}},"/device-logs":{"post":{"operationId":"DeviceLogsController_storeLog","parameters":[],"responses":{"201":{"description":""}},"tags":["DeviceLogs"]}},"/device-logs/{deviceSerial}":{"get":{"operationId":"DeviceLogsController_getDeviceLogs","parameters":[{"name":"deviceSerial","required":true,"in":"path","description":"Device serial","schema":{"type":"string"}},{"name":"severity","required":false,"in":"query","description":"Filter by log severity (info, warning, error)","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","description":"Start date (ISO format)","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","description":"End date (ISO format)","schema":{"type":"string"}}],"responses":{"200":{"description":"List of device logs."}},"summary":"Retrieve device logs","tags":["DeviceLogs"]}}},"info":{"title":"SmartReader API","description":"API for managing and monitoring SmartReader devices via MQTT and REST","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"x-api-key":{"type":"apiKey","in":"header","name":"x-api-key","description":"API key for authentication"}},"schemas":{"Object":{"type":"object","properties":{}}}}}